add Moore-notes
Saccharomyces cerevisiae, also known as the budding yeast, can be used
to study the cellular aging process as a model for aging of human
cells. Past studies have concluded that there is a strong connection
between the presence of superoxides, and cellular aging. This study
investigates how the presence of hydrogen peroxide on the outside of
the cell influences the amount of hydrogen peroxide inside of the cell.
The hydrogen peroxide acts as a stress inhibitor and the cell wants to
the keep a balance of between the levels of intercellular and
extracellular peroxide